Remove direct integration with NetworkManager (#192)
Now that https://github.com/openUC2/rpi-imswitch-os/pull/26 has been
merged, rpi-imswitch-os has relieved ImSwitch of the responsibilitiy to
know about the existence of NetworkManager (which ImSwitch had been
handling in a hacky way that involved installing `nmcli` in the ImSwitch
Docker container image and using `setuid` on `nmcli`). Thus, for now we
will rely on the [device-admin
app](https://github.com/openUC2/device-admin) as the GUI for users to
manage the Wi-Fi network, and we can link to it from ImSwitch. In the
farther future, I can modify the device-admin app to expose a JSON HTTP
API for programmatically doing things which currently can only be done
through device-admin's web GUI; then ImSwitch would be provide its own
GUI for the functionalities managed by device-admin, if/when we want to
do the work needed for that.

This PR also makes some small changes/fixes, namely:
- This PR changes ImSwitch's default WebSocket port configuration from
8002 to 8001 (@beniroquai lmk if you'd like me to just delete the
WebSocket port setting entirely)
- This PR disables Jupyter's authentication mechanism entirely, so that
a token no longer needs to be provided in the URL to access Jupyter.
(@beniroquai lmk if this would be a problem)
- This PR speeds up the ctr img build process slightly when only Python
files are edited (and `pyproject.toml` isn't edited), from ~4 min to ~3
min, by improving the use of container image layer caching. Edit: more
importantly, it significantly reduces the amount of time (and disk
storage) needed for upgrading to new versions of the ImSwitch Docker
container image when only Python files were modified, by removing the
need to re-download pip packages which hadn't changed between the
updates.

This PR is tracked on Notion at
https://www.notion.so/Remove-ImSwitch-NetworkManager-integration-2724e612c78a8060ad78e6375d241da7?source=copy_link